A number of cleanups recommended by Clang

+++ physics/solar_system_body.hpp
@@ -634,10 +634,12 @@ template<typename Frame>
 template<typename Message>
 void SolarSystem<Frame>::CheckFrame(Message const& message) {
   if (message.has_solar_system_frame()) {
-    CHECK_EQ(Frame::tag, message.solar_system_frame());
+    CHECK_EQ(static_cast<int>(Frame::tag),
+             static_cast<int>(message.solar_system_frame()));
   }
   if (message.has_plugin_frame()) {
-    CHECK_EQ(Frame::tag, message.plugin_frame());
+    CHECK_EQ(static_cast<int>(Frame::tag),
+             static_cast<int>(message.plugin_frame()));
   }
 }
